+++ lib/src/analyzers/lint_analyzer/reporters/reporter_factory.dart
@@ -1,12 +1,12 @@
 import 'dart:io';
 
-import '../../../config_builder/models/config.dart';
 import '../../../reporters/models/code_climate_reporter.dart';
 import '../../../reporters/models/console_reporter.dart';
 import '../../../reporters/models/github_reporter.dart';
 import '../../../reporters/models/html_reporter.dart';
 import '../../../reporters/models/json_reporter.dart';
 import '../../../reporters/models/reporter.dart';
+import '../lint_config.dart';
 import 'reporters_list/code_climate/lint_code_climate_reporter.dart';
 import 'reporters_list/console/lint_console_reporter.dart';
 import 'reporters_list/github/lint_github_reporter.dart';
@@ -14,7 +14,7 @@ import 'reporters_list/html/lint_html_reporter.dart';
 import 'reporters_list/json/lint_json_reporter.dart';
 
 final _implementedReports = <String,
-    Reporter Function(IOSink output, Config config, String reportFolder)>{
+    Reporter Function(IOSink output, LintConfig config, String reportFolder)>{
   ConsoleReporter.id: (output, _, __) => LintConsoleReporter(output),
   ConsoleReporter.verboseId: (output, _, __) =>
       LintConsoleReporter(output, reportAll: true),
@@ -34,7 +34,7 @@ final _implementedReports = <String,
 Reporter? reporter({
   required String name,
   required IOSink output,
-  required Config config,
+  required LintConfig config,
   required String reportFolder,
 }) {
   final constructor = _implementedReports[name];