+++ lib/src/analyzers/unused_files_analyzer/unused_files_analysis_config.dart
@@ -1,0 +1,11 @@
+import 'package:glob/glob.dart';
+
+class UnusedFilesAnalysisConfig {
+  final Iterable<Glob> globalExcludes;
+  final Iterable<Glob> analyzerExcludedPatterns;
+
+  const UnusedFilesAnalysisConfig(
+    this.globalExcludes,
+    this.analyzerExcludedPatterns,
+  );
+}