+++ lib/src/config_builder/config_builder.dart
@@ -1,35 +1,30 @@
 import '../analyzers/lint_analyzer/anti_patterns/patterns_factory.dart';
+import '../analyzers/lint_analyzer/lint_analysis_config.dart';
 import '../analyzers/lint_analyzer/lint_config.dart';
 import '../analyzers/lint_analyzer/metrics/metrics_factory.dart';
 import '../analyzers/lint_analyzer/metrics/models/metric.dart';
 import '../analyzers/lint_analyzer/models/entity_type.dart';
 import '../analyzers/lint_analyzer/rules/rules_factory.dart';
+import '../analyzers/unused_files_analyzer/unused_files_analysis_config.dart';
 import '../analyzers/unused_files_analyzer/unused_files_config.dart';
 import '../cli/models/parsed_arguments.dart';
 import '../utils/exclude_utils.dart';
 import 'models/analysis_options.dart';
-import 'models/config.dart';
 
 class ConfigBuilder {
-  static Config getConfig(
-    AnalysisOptions options, [
-    ParsedArguments? arguments,
-  ]) =>
-      arguments != null
-          ? Config.fromAnalysisOptions(options)
-              .merge(Config.fromArgs(arguments))
-          : Config.fromAnalysisOptions(options);
+  static LintConfig getLintConfigFromOptions(AnalysisOptions options) =>
+      LintConfig.fromAnalysisOptions(options);
 
-  static Config getConfigFromArgs(ParsedArguments arguments) =>
-      Config.fromArgs(arguments);
+  static LintConfig getLintConfigFromArgs(ParsedArguments arguments) =>
+      LintConfig.fromArgs(arguments);
 
-  static LintConfig getLintConfig(
-    Config config,
+  static LintAnalysisConfig getLintAnalysisConfig(
+    LintConfig config,
     String rootPath, {
     Iterable<Metric<num>>? classMetrics,
     Iterable<Metric<num>>? functionMetrics,
   }) =>
-      LintConfig(
+      LintAnalysisConfig(
         prepareExcludes(config.excludePatterns, rootPath),
         getRulesById(config.rules),
         getPatternsById(config.antiPatterns),
@@ -47,9 +42,22 @@ class ConfigBuilder {
         config.metrics,
       );
 
-  static UnusedFilesConfig getUnusedFilesConfig(
-    String rootPath,
+  static UnusedFilesConfig getUnusedFilesConfigFromArgs(
     Iterable<String> excludePatterns,
   ) =>
-      UnusedFilesConfig(prepareExcludes(excludePatterns, rootPath));
+      UnusedFilesConfig.fromArgs(excludePatterns);
+
+  static UnusedFilesConfig getUnusedFilesConfigFromOption(
+    AnalysisOptions options,
+  ) =>
+      UnusedFilesConfig.fromAnalysisOptions(options);
+
+  static UnusedFilesAnalysisConfig getUnusedFilesConfig(
+    UnusedFilesConfig config,
+    String rootPath,
+  ) =>
+      UnusedFilesAnalysisConfig(
+        prepareExcludes(config.excludePatterns, rootPath),
+        prepareNormalizedExcludes(config.analyzerExcludePatterns, rootPath),
+      );
 }