+++ test/analyzers/unused_files_analyzer/unused_files_analyzer_test.dart
@@ -1,11 +1,9 @@
 @TestOn('vm')
 import 'dart:io';
 
-import 'package:dart_code_metrics/src/analyzers/lint_analyzer/models/lint_file_report.dart';
 import 'package:dart_code_metrics/src/analyzers/unused_files_analyzer/reporters/reporters_list/console/unused_files_console_reporter.dart';
 import 'package:dart_code_metrics/src/analyzers/unused_files_analyzer/unused_files_analyzer.dart';
 import 'package:dart_code_metrics/src/analyzers/unused_files_analyzer/unused_files_config.dart';
-import 'package:dart_code_metrics/src/config_builder/config_builder.dart';
 import 'package:path/path.dart';
 import 'package:test/test.dart';
 
@@ -25,13 +23,12 @@ void main() {
       ];
 
       test('should analyze files', () async {
-        final config = _createConfig(rootDirectory);
+        final config = _createConfig(analyzerExcludePatterns: analyzerExcludes);
 
         final result = await analyzer.runCliAnalysis(
           folders,
           rootDirectory,
           config,
-          analyzerExcludes,
         );
 
         final report = result.single.relativePath;
@@ -49,14 +46,10 @@ void main() {
   );
 }
 
-UnusedFilesConfig _createConfig(
-  String rootDirectory, {
-  Iterable<String> excludePatterns = const [],
+UnusedFilesConfig _createConfig({
+  Iterable<String> analyzerExcludePatterns = const [],
 }) =>
-    ConfigBuilder.getUnusedFilesConfig(rootDirectory, excludePatterns);
-
-LintFileReport reportForFile(
-  Iterable<LintFileReport> reports,
-  String fileName,
-) =>
-    reports.firstWhere((report) => report.relativePath.endsWith(fileName));
+    UnusedFilesConfig(
+      excludePatterns: const [],
+      analyzerExcludePatterns: analyzerExcludePatterns,
+    );